requirements.txt
fastapi
uvicorn
openai
requests
python-dotenv
Jinja2
beautifulsoup4




